Add patches that allows to build the Linaro GCC the OE-Core way
The 64bithack.patch causes the GCC to install libssp and libstdc++-v3 into
lib as opposed to lib64 on X86_64 Linux targets. Usually the libstdc++ will
be placed in lib64 during the install step. Because that's where 64bit libs
go for linux targets according to the FHS. If you don't like this but want
to use linux as a target you have to patch the toolchain which is what the
64bithack.patch does.

The gcc-4.3.1-ARCH_FLAGS_FOR_TARGET.patch changes the compiler to respect OEs
ARCH_FLAGS_FOR_TARGET environment variable when building libgcc and friends.
This variable contains things like -march= and -mfloat-abi=. This change
fixes a build failure when using the Linaro GCC in conjunction with a hard
float configuration.

Signed-off-by: Ken Werner <ken.werner@linaro.org>