gcc: use fix-g++-sysroot.patch from gcc-4.7 for gcc-4.8
The c++ include paths for the configured g++ on the target are missing
a leading '/' which means various headers files are not found when
compiling c++ programs.

This commit carries forward the existing patch from gcc-4.7 which
ensures that the gcc-runtime configure stage will not drop the leading
'/' when --with-sysroot is specified.

Fixes https://bugs.launchpad.net/linaro-oe/+bug/1192953.

Signed-off-by: Andrew McDermott <andrew.mcdermott@linaro.org>