auto-serial-console: fix auto-getty hardcodes device names (LP: #1233786)
The auto-getty script is very badly behaved, as it assumes that the
console on which to run getty must be called tty*.
While this is true most of the time, this doesn't fly at all when using a
paravirtualized console (the standard hvc console, for example). You end
up with a booted guest and no way to interact with it.

The attached patch tries to work around this issue and parse
/etc/securetty to find out if the device passed on the kernel command line
has been vetted for root use. If so, there is no reason to prevent getty
from running on this console.

It also fix a very silly bug where GETTY_ARGS will forever expand each
time getty is re-spawned.

Change-Id: I6109d48bca13850cf90a37f11fd9aa24ce2dba43
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>