busybox: enable CONFIG_FEATURE_TAR_NOPRESERVE_TIME
When running JFFS2/UBIFS images in LAVA (e.g. from NAND), the images
cannot be modified upfront by LAVA to add the tests, so the tests need
to be untar'd upon boot from the target. 'tar -m' is required to avoid
messing up with the files mtime, otherwise we might get funny results
depending on the device clock versus the stored mtime in the
tarballs...

Change-Id: I5ce637b77b132a1bf6f07c610003186330a97a9e
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>