busybox: Background udhcpc if lease is not immediately obtained
The default busybox udhcpc configuration sets the --now option
"Exit with failure if lease is not immediately obtained".

The LAVA lab has a delay when servicing DHCP requests for boards
running OE - see bug 1239820.

This patch sets the -b option instead of -n, so udhcpc
will go into the background if the lease is not immediately obtained,
allowing networking tests to succeed when run in the LAVA lab.

Change-Id: I1a3ef36638507a78b8a5ffa95d6f2e1b58f5427e
Signed-off-by: Kim Phillips <kim.phillips@linaro.org>