meta-linaro: openjdk-8: add runtime recommendation on liberation-fonts
This is a temporary fix to ensure that there are some workable fonts for
OpenJDK-8 on OE.  Anything consuming or depending on AWT graphics
currently produces a NullPointerException. This has (possibly) changed
in 13.09 => 13.10 because we have merged to a later version of the tip
of x86 development and there are configure-based changes around font
configuration in general.  These changes need to now be understood and
considered.

Change-Id: I738b81468585eb2bae5d465bcf6489b927404784
Signed-off-by: Andrew McDermott <andrew.mcdermott@linaro.org>