linaro-image-lng: uniform hostname on all machines
This patch overrides the machine-dependent OE default hostname,
which is the 'machine name' for which the image was built.
The new default hostname is 'lng-test-target', and is uniform
across all archtectures.

This reduces the number of root login prompts which LAVA must
compare against when interacting with the system console,
(per request from the LAVA team).

At the same time it may be helpful in identifying machines which are
running on LNG OE filesystems in scenarios where users are accessing
the test machines interactively and/or remotely.

Change-Id: Ife86ae1907bdd48ef80eef87e0a9b349a272a099
Signed-off-by: Gary S. Robertson <gary.robertson@linaro.org>