odp: add support for ptest
ODP has a comprehensive set of API validation tests, so build and
install these in order for them to be run on the target platform.
The ptest framework is used to do this, allowing some previously
unused PACKAGECONFIG options to be tidied up.

To enable this feature "ptest" must be added to DISTRO_FEATURES.

Change-Id: I1cc9beb4876ae2ec1fe4e4d05b841cee5bf779e0
Signed-off-by: Stuart Haslam <stuart.haslam@linaro.org>