linaro-image-lamp: remove grub when building for armv7a
Grub git was updated with checks for softfloat and thumb which both clash with our target CFLAGS. Grub makes it impossible to specify CFLAGS *only* for the userspace utils. The toolchain defaults to softfloats, so zapping CFLAGS leads to configure failing to link to glibc...

Anyway, since we don't care about armv7a/grub, just don't build it.

Change-Id: I83ed42720776d4d54a08cd7bbdf8cea1d4ab9faf
Signed-off-by: Koen Kooi <koen.kooi@linaro.org>