meta-ilp32: perl: undefine libm_lib_version for ILP32
Recent versions of GLIBC got rid of _LIB_VERSION, and said
versions are the ones on which ILP32 is going to land, so
we can safely say that for ILP32 will run on that recent
GLIBC. Besides, said condition is only used for BSD and
Cygwin, so no harm done.

Generally, this shouldn't be needed, but we bypass Configure
and predefine config.sh in the recipe, so this just ensures
that the config file is correct for ILP32.

Change-Id: I11f6f8db584759492aa3a0e24867af0e396e1f72
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>