external-arm-toolchain: libc.so linker script should use relative path
libc.so linker script should use relative paths to not clash with host libs.
Using absolute paths in libc.so is fine with standalone toolchain or when
used inside a rootfs. But when used inside OE SDK for cross-compilation,
the absolute path confuses compiler with libs from host environment. In the
past, Linaro prebuilt toolchains always had relative paths inside libc.so
linker script. Adjust it the same for the Arm prebuilt toolchain.

Signed-off-by: Denys Dmytriyenko <denys@ti.com>
Change-Id: I8faaddcc11da7cfb9fb29217084cbc0d34dcc270